pos-edi/services/pos-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.5
httpx==0.27.0
pydantic==2.8.2
kafka-python==2.0.2
